requirements.txt
duckdb
pandas
dbt-duckdb
logging
pandas as pd
matplotlib.pyplot as plt
seaborn as sns
datetime
